requirements.txt
certifi==2025.1.31
charset-normalizer==2.1.1
idna==3.10
mysql-connector-python==8.0.33
protobuf==3.20.3
python-dotenv==1.1.0
requests==2.28.1
urllib3==1.26.20
